requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
idna==3.7
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.5
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
